deps: update V8 to 6.9.274

+++ deps/v8/src/string-search.h
@@ -567,9 +567,9 @@ int SearchString(Isolate* isolate,
 // and pattern as vectors before calling SearchString. Used from the
 // StringIndexOf builtin.
 template <typename SubjectChar, typename PatternChar>
-int SearchStringRaw(Isolate* isolate, const SubjectChar* subject_ptr,
-                    int subject_length, const PatternChar* pattern_ptr,
-                    int pattern_length, int start_index) {
+intptr_t SearchStringRaw(Isolate* isolate, const SubjectChar* subject_ptr,
+                         int subject_length, const PatternChar* pattern_ptr,
+                         int pattern_length, int start_index) {
   DisallowHeapAllocation no_gc;
   Vector<const SubjectChar> subject(subject_ptr, subject_length);
   Vector<const PatternChar> pattern(pattern_ptr, pattern_length);
